requirements.txt
python-nmap
pexpect
ptyprocess